♻️ Split explorer & indexer
explorer only run with: HYLE_RUN_INDEXER=false
indexer only run with: HYLE_RUN_EXPLORER=false

can also HYLE_RUN_REST_SERVER=false for indexer, but usefull for health
check
