+++ src/entrypoint.rs
@@ -4,6 +4,7 @@ use crate::{
     bus::{metrics::BusMetrics, SharedMessageBus},
     consensus::Consensus,
     data_availability::DataAvailability,
+    explorer::Explorer,
     genesis::Genesis,
     indexer::Indexer,
     mempool::Mempool,
@@ -126,7 +127,7 @@ pub fn welcome_message(conf: &conf::Conf) {
    ██║  ██║╚██╗ ██╔╝██║     ██║         {validator_details}
    ███████║ ╚████╔╝ ██║     ██║       {check_p2p} p2p::{p2p_port} | {check_http} http::{http_port} | {check_tcp} tcp::{tcp_port} | ◆ da::{da_port}
    ██╔══██║  ╚██╔╝  ██║     ██║     
-   ██║  ██║   ██║   ███████╗██║     {check_indexer} indexer {database_url}
+   ██║  ██║   ██║   ███████╗██║     {check_indexer} indexer {check_explorer} explorer db: {database_url}
    ╚═╝  ╚═╝   ╚═╝   ╚══════╝╚═╝     ∎ {data_directory}
  
    Minimal, yet sufficient. Hope You Like It.
@@ -149,6 +150,7 @@ pub fn welcome_message(conf: &conf::Conf) {
         tcp_port = conf.tcp_server_port,
         da_port = conf.da_public_address,
         check_indexer = check_or_cross(conf.run_indexer),
+        check_explorer = check_or_cross(conf.run_explorer),
         database_url = if conf.run_indexer {
             format!("↯ {}", mask_postgres_uri(conf.database_url.as_str()))
         } else {
@@ -300,6 +302,12 @@ async fn common_main(
             .await?;
     }
 
+    if config.run_explorer {
+        handler
+            .build_module::<Explorer>((config.clone(), build_api_ctx.clone()))
+            .await?;
+    }
+
     if config.p2p.mode != conf::P2pMode::None {
         let ctx = SharedRunContext {
             config: config.clone(),
@@ -337,7 +345,7 @@ async fn common_main(
         }
 
         handler.build_module::<P2P>(ctx.clone()).await?;
-    } else {
+    } else if config.run_indexer {
         handler
             .build_module::<DAListener>(DAListenerConf {
                 data_directory: config.data_directory.clone(),
