+++ src/indexer/handler.rs
@@ -1,4 +1,4 @@
-use super::api::*;
+use crate::explorer::api::*;
 use crate::model::*;
 use crate::node_state::module::NodeStateEvent;
 use anyhow::{bail, Context, Error, Result};
@@ -138,7 +138,7 @@ impl Indexer {
             return Ok(());
         }
 
-        let mut transaction = self.state.db.begin().await?;
+        let mut transaction = self.db.begin().await?;
 
         // Insert blocks into the database
         if !self.handler_store.blocks.is_empty() {
@@ -511,7 +511,7 @@ impl Indexer {
     }
 
     pub async fn handle_mempool_status_event(&mut self, event: MempoolStatusEvent) -> Result<()> {
-        let mut transaction = self.state.db.begin().await?;
+        let mut transaction = self.db.begin().await?;
         match event {
             MempoolStatusEvent::WaitingDissemination {
                 parent_data_proposal_hash,
