+++ src/utils/conf_defaults.toml
@@ -24,6 +24,7 @@ tcp_server_port = 1414
 
 # Indexer
 run_indexer = true
+run_explorer = true
 database_url = "postgres://postgres:postgres@localhost:5432/postgres"
 da_read_from = "127.0.0.1:4141"
